refactor: Para evitar criação de pastas e arquivos desnecessários, definição do objeto de requisição do UseCase é definido no mesmo namespace
