--- Financeiro.Application/UseCases/R-2020/DTOs/EnviarR2020Requisicao.cs
@@ -1,4 +1,0 @@
-﻿namespace Financeiro.Application.UseCases.DespesasPorPeriodo.DTOs;
-public record EnviarR2020Requisicao(int Mes, int Ano)
-{
-}
